Remove examples from API/Components
Summary:
These examples used to be part of the UIExplorer app, which will be renamed to RNTester in #13606. These bits of code have not been treated as examples for a long time, and are actually part of our testing infra. Let's remove them altogether from the API/Components docs.
Closes https://github.com/facebook/react-native/pull/13664

Differential Revision: D4960725

Pulled By: hramos

fbshipit-source-id: a4ec19b81d992aff89c0c93e35a1b80fbbd5e454